update whitepaper to v1.1 (#22)

+++ contracts/libs/Timebased.sol
@@ -3,27 +3,27 @@ pragma solidity ^0.5.0;
 import "openzeppelin-solidity/contracts/math/SafeMath.sol";
 
 contract Timebased {
-	using SafeMath for uint;
+	using SafeMath for uint256;
 	struct BaseTime {
-		uint time;
-		uint blockHeight;
+		uint256 time;
+		uint256 blockHeight;
 	}
 	BaseTime internal baseTime;
-	uint internal secondsPerBlock = 15;
+	uint256 internal secondsPerBlock = 15;
 
 	constructor() public {
 		// solium-disable-next-line security/no-block-members
 		baseTime = BaseTime(now, block.number);
 	}
 
-	function _setSecondsPerBlock(uint _sec) internal {
+	function _setSecondsPerBlock(uint256 _sec) internal {
 		secondsPerBlock = _sec;
 	}
 
-	function timestamp() internal view returns (uint) {
-		uint diff = block.number - baseTime.blockHeight;
-		uint sec = diff.div(secondsPerBlock);
-		uint _now = baseTime.time.add(sec);
+	function timestamp() internal view returns (uint256) {
+		uint256 diff = block.number - baseTime.blockHeight;
+		uint256 sec = diff.div(secondsPerBlock);
+		uint256 _now = baseTime.time.add(sec);
 		return _now;
 	}
 }
